try to fix sonar
